Refactor landing page components and styles; add new animated stats and CTA section
