+++ package.json
@@ -31,6 +31,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-element-to-jsx-string": "^17.0.0",
+    "recharts": "^2.15.1",
     "tago": "file:",
     "tailwind-merge": "^3.0.1",
     "zod": "^3.24.1"
