+++ src/app/page.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import Head from 'next/head';
 import Navbar from '@/components/Navbar';
-import HeroSection from '@/components/HeroSection';
-import FeaturedDesigns from '@/components/FeaturedDesigns';
-import StatsSection from '@/components/StatsSection';
-import FeaturesSection from '@/components/FeaturesSection';
-import TestimonialsSection from '@/components/TestimonialsSection';
-import CTASection from '@/components/CTASection';
+import HeroSection from '@/components/Landing Page/HeroSection';
+import FeaturedDesigns from '@/components/Landing Page/FeaturedDesigns';
+import StatsSection from '@/components/Landing Page/StatsSection';
+import FeaturesSection from '@/components/Landing Page/FeaturesSection';
+import TestimonialsSection from '@/components/Landing Page/TestimonialsSection';
+import CTASection from '@/components/Landing Page/CTASection';
 import Footer from '@/components/Footer';
-import GlobalStyles from '@/components/GlobalStyles';
+import GlobalStyles from '@/components/Landing Page/GlobalStyles';
 import ScrollNavWrapper from '@/components/ScrollNavWrapper';
 
 const LandingPage = () => {
@@ -23,8 +23,7 @@ const LandingPage = () => {
       <div className="min-h-screen bg-black text-white overflow-hidden">
         <ScrollNavWrapper>
           <Navbar />
-        </ScrollNavWrapper>
-        
+        </ScrollNavWrapper> 
         <HeroSection />
         <FeaturedDesigns />
         <StatsSection />
