+++ src/components/Landing Page/AnimatedStatsSection.tsx
