+++ src/components/Landing Page/GlobalStyles.tsx
